+++ app/helpers/private_helper.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+
+# Provate Helper
+module PrivateHelper
+end