+++ config/application.rb
@@ -26,5 +26,6 @@ module EasyPostIaBackend
     #
     # config.time_zone = "Central Time (US & Canada)"
     # config.eager_load_paths << Rails.root.join("extras")
+    config.web_console.whitelisted_ips = '192.168.65.1'
   end
 end