+++ CHANGELOG.md
@@ -2,6 +2,21 @@
 
 This document lists breaking changes for each major release.
 
+## v0.0.2 (21/09/2024)
+
+Feature: KAN-26, KAN-2, KAN-28, Create login, logout and signup.
+
+- Create configuration to Cors, to console and debug.
+- Configure devise with JWT and strategy to revocate.
+- Create helpers to auth.
+- Update .gitignore.
+- Update Rbuocop rules.
+- Add initial swagger.
+- Create initial routes, controllers to auth.
+- Add and update config to test with Rspec, FactoryBot, ShouldMatchers, Devise test, coverage.
+- Add test to models, controllers, helpers, etc.
+- Update Gemfile to new gems to clean database in the tests.
+
 ## v0.0.1 (20/09/2024)
 
 Feature: KAN-13, Config project and initial projects